chore: move Device to own module
