+++ src/network_puzzles/gui/networkpuzzles.kv
@@ -3,6 +3,7 @@
 #:import session network_puzzles.session
 #:include network_puzzles/gui/base.kv
 #:include network_puzzles/gui/buttons.kv
+#:include network_puzzles/gui/device.kv
 #:include network_puzzles/gui/labels.kv
 #:include network_puzzles/gui/layouts.kv
 #:include network_puzzles/gui/popups.kv
